+++ service/tier1.go
@@ -828,8 +828,10 @@ func tier1ResponseHandler(ctx context.Context, mut *sync.Mutex, logger *zap.Logg
 			return connect.NewError(connect.CodeUnavailable, err)
 		}
 
-		stats.RecordMessageSent(isData, egressBytes)
-		metering.AddEgressBytes(ctx, egressBytes)
+		if isData {
+			stats.RecordDataSent(egressBytes)
+			metering.AddEgressBytes(ctx, egressBytes)
+		}
 		metericsSender.Send(ctx, userID, apiKeyID, ip, userMeta, outputModuleHash, "sf.substreams.rpc.v2/Blocks")
 		return nil
 	}